+++ client/src/App.tsx
@@ -1,63 +1,19 @@
-import React, {useState} from 'react';
+import {Routes, Route} from 'react-router-dom';
+import Home from './pages/Home/Home';
+import About from './pages/About/About';
+import NavBar from './components/NavBar/NavBar';
 
-interface TeamResponse {
-  createdAt: string;
-  id: number;
-  name: string;
-  updatedAt: string;
-}
-
-function App() {
-  const [inputText, setInputText] = useState('');
-  const [responseData, setResponseData] = useState<TeamResponse | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  const fetchData = async () => {
-    try {
-      setLoading(true);
-
-      const response = await fetch(`/create/${inputText}`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        // body: JSON.stringify({ input: inputText }),
-      });
-
-      const result = await response.json();
-      setResponseData(result.data);
-
-      setLoading(false);
-    } catch (error) {
-      console.error('Error fetching data:', error);
-      setLoading(false);
-    }
-  };
+const App = () => {
+  
 
   return (
-    <div>
-      <label>
-        Enter Name:
-        <input
-          type="text"
-          value={inputText}
-          onChange={(e) => setInputText(e.target.value)}
-        />
-      </label>
-      <button onClick={fetchData} disabled={loading}>
-        Fetch Data
-      </button>
-      {loading && <p>Loading...</p>}
-      {responseData && (
-        <div>
-          <h2>Response Data:</h2>
-          <p>created at: {responseData.createdAt}</p>
-          <p>id: {responseData.id}</p>
-          <p>name: {responseData.name}</p>
-          <p>updated at: {responseData.updatedAt}</p>
-        </div>
-      )}
-    </div>
+    <>
+      <NavBar/>
+      <Routes>
+        <Route path="/" element={<Home/>} />
+        <Route path="/about" element={<About/>} />
+      </Routes>
+    </>
   );
 }
