+++ client/src/components/NavBar/NavBar.tsx
@@ -1,0 +1,20 @@
+import { Link } from 'react-router-dom';
+import styles from './NavBar.module.css';
+
+const NavBar = () => {
+    return (
+    <nav className={styles.navbar}>
+        <div className="container mx-auto">
+            <div className="flex items-center justify-between">
+                <div className={styles['navbar-brand']}>Cascarita</div>
+                <div className="flex">
+                    <a href="/" className={styles['navbar-link']}>Home</a>
+                    <a href="/about" className={styles['navbar-link']}>About</a>
+                </div>
+            </div>
+        </div>
+    </nav>
+    );
+};
+
+export default NavBar;
