+++ client/src/pages/Home/Home.tsx
@@ -1,0 +1,9 @@
+const Home = () => {
+    return(
+        <div>
+            <p>Home Page</p>
+        </div>
+    );
+}
+
+export default Home;
